Add files via upload
The traveling salesman wants to visit n cities. For each city, he knows its
coordinates (x, y). Traveling between cities with coordinates (x1, y1) and (x2,
y2) takes him the amount of time proportional to their Euclidian distance
defined as follows:
In the above equation v is the constant velocity of the salesman. Additionally,
in each city the salesman has to spend a certain amount of time. The length of
this stay depends on the day of the week in which the salesman comes to this
city. We assume that the salesman visits the first city on Monday, and each
following city during the subsequent day of the week (the second city on
Tuesday, the third on Wednesday, etc.). The salesman wants to finish the
journey as quickly as possible, and, therefore, he also works on Saturdays and
Sundays.
Determine the route for a traveling salesman that minimizes the total time
spent on the trip and in the cities. If you propose a short route, he will have a
lot of time for visiting!